pyproject.toml
[project]
name = "mcp-servers"
version = "0.2.0"
description = "MCP server for EdTech / Lesson Kit - Cloud.ru / Evolution AI Agents"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.9.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "jinja2>=3.1.0",
    "requests>=2.32.5",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]
include = [
    "mcp_instance.py",
    "server.py",
    "templates/**/*.html",
]
